requirements.txt
pytest==7.1.1
selenium==4.1.3
allure-pytest==2.17.3